Adding ExposedPorts to hold the internet protocol
